Update 578. Get Highest Answer Rate Question.sql

+++ LeetCode/SQL/Medium/578. Get Highest Answer Rate Question.sql
@@ -2,7 +2,7 @@
 
          -- Approach 1. Using - GROUP BY and - FILTER -- 
 SELECT question_id, 
-    COUNT(*) FILTER(WHERE answer_id IS NOT NULL)::numeric 
+    COUNT(*) FILTER(WHERE answer_id IS NOT NULL)
       / COUNT(*) AS answer_rate
 FROM survey_log
 GROUP BY question_id
@@ -13,7 +13,7 @@ LIMIT 1;
 
          -- Approach 2. Using - GROUP BY with - HAVING -- 
 SELECT question_id, 
-    COUNT(*) FILTER(WHERE answer_id IS NOT NULL)::numeric 
+    COUNT(*) FILTER(WHERE answer_id IS NOT NULL)
       / COUNT(*) AS answer_rate
 FROM survey_log
 GROUP BY question_id
